+++ CHANGELOG.md
@@ -5,6 +5,9 @@ FlatLaf Change Log
 
 #### New features
 
+- Added API to register packages or folders where FlatLaf searches for
+  application specific properties files with custom UI defaults (see
+  `FlatLaf.registerCustomDefaultsSource(...)` methods).
 - Extras: `FlatSVGIcon` now allows specifying `ClassLoader` that is used to load
   SVG file. (issue #163)